refactor(app): home new workspace project picker state
